Moved exclude_identification_keys to under peer_forwarder config
