+++ data-prepper-core/src/main/java/org/opensearch/dataprepper/core/parser/model/DataPrepperConfiguration.java
@@ -22,12 +22,10 @@ import org.opensearch.dataprepper.plugin.ExtensionsConfiguration;
 
 import java.time.Duration;
 import java.util.Collections;
-import java.util.stream.Collectors;
 import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.Objects;
 
 /**
@@ -58,7 +56,6 @@ public class DataPrepperConfiguration implements ExtensionsConfiguration, EventC
     private Duration processorShutdownTimeout;
     private Duration sinkShutdownTimeout;
     private PipelineExtensions pipelineExtensions;
-    private List<Set<String>> excludeIdentificationKeys;
 
     public static final DataPrepperConfiguration DEFAULT_CONFIG = new DataPrepperConfiguration();
 
@@ -76,8 +73,6 @@ public class DataPrepperConfiguration implements ExtensionsConfiguration, EventC
             @JsonProperty("private_key_password")
             @JsonAlias("privateKeyPassword")
             final String privateKeyPassword,
-            @JsonProperty("exclude_identification_keys")
-            final List<Set<String>> excludeIdentificationKeys,
             @JsonProperty("server_port")
             @JsonAlias("serverPort")
             final String serverPort,
@@ -132,7 +127,6 @@ public class DataPrepperConfiguration implements ExtensionsConfiguration, EventC
             throw new IllegalArgumentException("sinkShutdownTimeout must be non-negative.");
         }
         this.pipelineExtensions = pipelineExtensions;
-        this.excludeIdentificationKeys = excludeIdentificationKeys;
     }
 
     public int getServerPort() {
@@ -141,13 +135,6 @@ public class DataPrepperConfiguration implements ExtensionsConfiguration, EventC
 
     public boolean ssl() {
         return ssl;
-    }
-
-    public Set<Set<String>> getExcludeIdentificationKeys() {
-        if (excludeIdentificationKeys == null) {
-            return null;
-        }
-        return excludeIdentificationKeys.stream().collect(Collectors.toSet());
     }
 
     public String getKeyStoreFilePath() {
