+++ data-prepper-core/src/test/java/org/opensearch/dataprepper/core/peerforwarder/PeerForwarder_ClientServerIT.java
@@ -519,6 +519,7 @@ class PeerForwarder_ClientServerIT {
                 200,
                 500,
                 1024,
+                null,
                 ssl,
                 sslCertificateFile,
                 sslKeyFile,
