+++ data-prepper-core/src/test/resources/valid_data_prepper_config.yml
@@ -1,3 +1,2 @@
 server_port: 5678
-exclude_identification_keys: [ ["key1"], ["key2", "key3"]]
 ssl: false
