+++ data-prepper-core/src/test/resources/valid_peer_forwarder_config.yml
@@ -4,6 +4,7 @@ client_timeout: 50
 server_thread_count: 100
 max_connection_count: 100
 max_pending_requests: 512
+exclude_identification_keys: [ ["key1"], ["key2", "key3"] ]
 ssl: false
 use_acm_certificate_for_ssl: false
 discovery_mode: static
